requirements.txt
flask
pylistenbrainz
pillow
requests
gunicorn
python-dotenv
